+++ scripts/criminals/Dialog.js
