+++ scripts/main.js
@@ -1,9 +1,9 @@
-import { getCriminals } from "./criminals/CriminalProvider.js";
+import { getCriminals } from "./criminals/criminalDataProvider.js";
 import { criminalList, criminalListButton } from "./criminals/CriminalList.js";
 import { ConvictionSelect } from "./convictions/ConvictionsSelect.js";
-import { getConvictions } from "./convictions/ConvictionsProvider.js";
+import { getConvictions } from "./convictions/convictionsDataProvider.js";
 import { NoteForm } from "./notes/NoteForm.js";
-import { witnessListButton } from "./witnesses/witnessList.js";
+import { witnessListButton } from "./witnesses/WitnessList.js";
 import { getWitnesses } from "./witnesses/witnessDataProvider.js";
 import { getNotes } from "./notes/noteDataProvider.js";
 import { noteListButton } from "./notes/NoteList.js";
